beacon committees fetch from archive
